+++ ansible_taskrunner/libs/click_extras/__init__.py
@@ -145,9 +145,19 @@ class ExtendCLI():
         nargs_ul_is_set = False
         nargs_ul_count = 0
         nargs_ul_count_max = 1
-        for cli_option, value in parameters.items():
+        prompt_if_missing = False
+        secure_input = False
+        option_type = None
+        value_from_env = None
+        for cli_option, value in parameters.items():            
             cli_option = Template(cli_option).safe_substitute(**self.available_vars)
-            value = Template(value).safe_substitute(**self.available_vars)
+            if isinstance(value, list):
+                value = [Template(v).safe_substitute(**self.available_vars) if isinstance(v, str) else v for v in value if isinstance(v, (str, int))]
+            elif isinstance(value, str):
+                value = Template(value).safe_substitute(**self.available_vars)
+            else:
+                logger.error('Skipping unsupported value type %s' % s)
+                continue
             opt_option = None
             opt_option_value = None
             cli_option_is_me = False
@@ -177,50 +187,80 @@ class ExtendCLI():
             cli_option_other_is_called = any(cli_option_other_matches) and len(cli_option_other_matches) == len_cli_option_other
             # Check for option with variadic arguments (nargs=*)
             # There can only be one!
-            if isinstance(value, list):
+            # Also, make sure we're not treating click.Choice types
+            if isinstance(value, list) and '|choice' not in cli_option:
                 if len(value) == 1:
                     nargs_ul_is_set = True
                     nargs_ul_count += 1
                     value = str(value[0])
                     numargs = -1
                 elif len(value) > 1:
-                    numargs = value[1]
+                    if isinstance(value[1], int):
+                        numargs = value[1] # e.g. [myvalue,2] would mean the option for myvalue requires 2 args
                     value = str(value[0])
+            # Account for click.Choice types
+            elif isinstance(value, list):
+                option_type = 'choice'
             if self.help_msg_map.get(cli_option):
                 option_help = self.help_msg_map[cli_option]
             else:
                 option_help = value
             if '|' in cli_option:
-                option_string = cli_option.split('|')
-                first_option = option_string[0].strip()
-                second_option = option_string[1].strip()
+                option_strings = cli_option.split('|')
+                first_option = option_strings[0].strip()
+                second_option = option_strings[1].strip()
+                if len(option_strings) > 2:
+                    option_tags = [o.strip() for o in option_strings[2:]]
+                    if 'prompt' in option_tags:
+                        prompt_if_missing = True
+                        secure_input = False
+                    if 'sprompt' in option_tags:
+                        prompt_if_missing = True
+                        secure_input = True
+                    if 'env' in option_tags:
+                        value_from_env = value
                 numargs = 1 if numargs < 1 else numargs
                 if opt_option:
-                    option = click.option(first_option, second_option, value, 
-                    type=str, nargs=numargs, help=option_help, cls=NotRequiredIf,
-                    not_required_if=opt_option_value)
+                    if option_type == 'choice':
+                        option = click.option(first_option, second_option, 
+                        type=click.Choice(value), help=option_help, cls=NotRequiredIf,
+                        prompt=prompt_if_missing, hide_input=secure_input, not_required_if=opt_option_value,
+                        envvar=value_from_env)
+                    else:
+                        option = click.option(first_option, second_option, value, 
+                        type=str, nargs=numargs, help=option_help, cls=NotRequiredIf,
+                        prompt=prompt_if_missing, hide_input=secure_input, not_required_if=opt_option_value,
+                        envvar=value_from_env)
                 else:
                     if cli_option_is_me and cli_option_other_is_called:
                         _is_required = False
                     else:
                         _is_required = is_required
-                    option = click.option(first_option, second_option, 
-                    value, type=str, nargs=numargs, help=option_help, 
-                    required=_is_required)
+                    if option_type == 'choice':
+                        option = click.option(first_option, second_option, 
+                        type=click.Choice(value), help=option_help, 
+                        prompt=prompt_if_missing, hide_input=secure_input, required=_is_required,
+                        envvar=value_from_env)
+                    else:
+                        option = click.option(first_option, second_option, 
+                        value, type=str, nargs=numargs, help=option_help, 
+                        prompt=prompt_if_missing, hide_input=secure_input, required=_is_required,
+                        envvar=value_from_env)
             else:
                 if nargs_ul_is_set and \
                 not nargs_ul_count > nargs_ul_count_max:
                     option = click.argument(
                         cli_option, value, help=option_help, 
-                        nargs=numargs, required=is_required)
+                        nargs=numargs, prompt=prompt_if_missing, hide_input=secure_input, required=is_required,
+                        envvar=value_from_env)
                 else:
                     numargs = 1 if numargs < 1 else numargs
                     if opt_option:
                         option = click.option(
                             cli_option, value, cls=NotRequiredIf,
                             is_flag=True, default=False, 
-                            help=option_help, 
-                            not_required_if=opt_option_value)
+                            help=option_help, not_required_if=opt_option_value,
+                            envvar=value_from_env)
                     else:
                         if cli_option_is_me and cli_option_other_is_called:
                             _is_required = False
@@ -229,8 +269,16 @@ class ExtendCLI():
                         option = click.option(
                             cli_option, value, is_flag=True, 
                             default=False, help=option_help, 
-                            required=_is_required)
-            func = option(func)
+                            required=_is_required, envvar=value_from_env)
+            try:
+                func = option(func)
+            except Exception as e:
+                logger.error('I had a problem parsing your parameters, error was %s' % e)
+                continue
+            option_type = None
+            prompt_if_missing = False
+            secure_input = False
+            value_from_env = None
             nargs_ul_is_set = False
             numargs = 1
         return func